Reworked version for PIC 18f2520
This version is intended to work on a new display PCB which will have a PIC18F2520 replacing the old processor. Processors are not pin compatible, so the code will not work on the original processor on the original board ! Code was compiled using PIC18 compiler from Microsoft. Flashed to the board using PICKIT2 or ICD3.